change to scroll sensitivity

+++ app/src/main/java/com/github/charleslzq/pacsdemo/component/event/ImageDisplayEvent.kt
@@ -10,7 +10,7 @@ class ImageDisplayEvent {
     class ChangePlayStatus(layoutPosition: Int) : ImageCellEvent(layoutPosition)
     class PlayModeReset(layoutPosition: Int) : ImageCellEvent(layoutPosition)
     class IndexChange(layoutPosition: Int, val index: Int) : ImageCellEvent(layoutPosition)
-    class IndexScroll(layoutPosition: Int, val scroll: Int) : ImageCellEvent(layoutPosition)
+    class IndexScroll(layoutPosition: Int, val scroll: Float) : ImageCellEvent(layoutPosition)
     class LocationTranslate(layoutPosition: Int, val distanceX: Float, val distanceY: Float) : ImageCellEvent(layoutPosition)
     class StudyModeReset(layoutPosition: Int) : ImageCellEvent(layoutPosition)
     class DrawPath(layoutPosition: Int, val points: List<PointF>) : ImageCellEvent(layoutPosition)
